refactor:연관성이 높은 코드를 한 파일로 이동, 이에 맞춰 파일 이름 변경
