+++ src/view/viewConstans.js
@@ -1,3 +1,25 @@
+export const SELECTOR = {
+	APP: '#app',
+	CONTENTS_AREA: '#contents-area',
+	COIN_MENU: '#vending-machine-manage-menu',
+	PRODUCT_MENU: '#product-add-menu',
+	PURCHASE_MENU: '#product-purchase-menu',
+	COIN_CHARGE_INPUT: '#vending-machine-charge-input',
+	COIN_CHARGE_BUTTON: '#vending-machine-charge-button',
+	COIN_500: '#vending-machine-coin-500-quantity',
+	COIN_100: '#vending-machine-coin-100-quantity',
+	COIN_50: '#vending-machine-coin-50-quantity',
+	COIN_10: '#vending-machine-coin-10-quantity',
+	PRODUCT_NAME_INPUT: '#product-name-input',
+	PRODUCT_PRICE_INPUT: '#product-price-input',
+	PRODUCT_QUANTITY_INPUT: '#product-quantity-input',
+	PRODUCT_ADD_BUTTON: '#product-add-button',
+	PURCHASE_CHARGE_INPUT: '#charge-input',
+	PURCHASE_CHARGE_AMOUNT: '#charge-amount',
+	PURCHASE_CHARGE_BUTTON: '#charge-button',
+	PURCHASE_ITEM_BUTTON: '.purchase-button',
+	PURCHASE_ITEM_QUANTITY: '.product-purchase-quantity',
+};
 export const TEMPLATES = {
 	HEADER: `
 	<h1>자판기</h1>
